added permission based on user roles

+++ src/pages/AdminOrg/Organization.js
@@ -11,6 +11,7 @@ import ServiceApi from "../../services/Service";
 import AddOrganization from "./AddOrganization";
 import { useDispatch } from "react-redux";
 import { fetchOrg } from "../../action";
+import { getCookies } from "../../utils/Utility";
 
 const { confirm } = Modal;
 
@@ -26,6 +27,8 @@ const Organization = function ({ currentLang }) {
 
   const { t } = useTranslation();
   const dispatch = useDispatch();
+  const checkAdmin = getCookies("user_token")?.user?.roles?.find(item=>item.calendarId==="CULTURE_OUTAOUAIS")
+
 
   const eventTableHeader = [
     {
@@ -71,11 +74,12 @@ const Organization = function ({ currentLang }) {
       key: "hasDependency",
       width:100,
       render: (e, record) => (
+        checkAdmin && (checkAdmin.role === "ADMIN" || checkAdmin.role === "SUPER_ADMIN")?
         <DeleteOutlined
           style={{fontSize:"23px"}}
           onClick={(event) => handleDelete(record, event)}
          
-        />
+        />:<></>
       ),
     },
   ];
@@ -221,7 +225,8 @@ const Organization = function ({ currentLang }) {
                 return {
                   onClick: (event) => {
                     event.stopPropagation()
-                    navigate(`/admin/add-organization/?id=${record.uuid}`);
+                    if(getCookies("user_token")?.user?.id===record.creator?.userId ||(checkAdmin && (checkAdmin.role === "EDITOR" || checkAdmin.role === "ADMIN" || checkAdmin.role === "SUPER_ADMIN")))
+                     navigate(`/admin/add-organization/?id=${record.uuid}`);
                     // setSelectedProduct(record);
                   }, // click row
                 };
